first steps towards splitting tests
